separated Filter into separate component
