+++ src/components/ListTitle.css
@@ -1,0 +1,4 @@
+form {
+	display: flex;
+	justify-content: center;
+}
