+++ src/components/ListTitle.tsx
@@ -1,6 +1,7 @@
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faPencil } from '@fortawesome/free-solid-svg-icons';
 import '../views/List.css';
+import './ListTitle.css';
 
 const ListTitle = ({ editName, isDisabled, updateListName, listName }) => {
 	return (
